Refatora e reorganiza classes de teste
Refatora e move as classes `GuidTypeHandlerForTests` e `SqlServiceTests` para novos namespaces e arquivos. Renomeia `SqlServiceTests` para `SqlHelperTests`. Ajusta importações de namespaces e mantém a estrutura e lógica dos testes unitários, garantindo a funcionalidade do `SqlHelper`.
